[Ant4C] added buffer_release_inner_buffers function.
Code in this function early located at buffer_release_with_inner_buffers
function.

Added common_get_attributes_and_arguments_for_task function that contain
code early located in several locations at units echo, exec, project and
property.

Replaced loops while(1) with for(;;) at date_time unit.

Replaced function interpreter_get_value_from_quote with string_un_quote.

Function property_get_attributes_and_arguments_for_task not set default
values of task property, it checked at property_evaluate_task function.
If some value not set at xml script code it set with default values
here.

Added support to function string_un_quote work with arguments of
functions from interpreter unit and strings that contain only quote
symbols.

Renamed test echo_evaluate_task to interpreter_evaluate_task.

Resolved compile warning at tests_exec - casting from uint8_t to bool
may generate performance issue.

Removed exec_get_attributes_and_arguments_for_task,
echo_get_attributes_and_arguments_for_task,
project_get_attributes_and_arguments_for_task and
property_get_attributes_and_arguments_for_task.

Resolved memory leak at project_load_from_build_file test.

+++ tests_text_encoding.cpp
@@ -626,7 +626,8 @@ TEST(TestTextEncoding_, text_encoding_UTF8_to_UTF16LE)
 	//
 	uint32_t returned_result = 0;
 
-	for (expected_return = 0x0; expected_return < buffer_size(&input_UTF8); ++expected_return, ++returned_result)
+	for (expected_return = 0x0; expected_return < (uint32_t)buffer_size(&input_UTF8);
+		 ++expected_return, ++returned_result)
 	{
 		const uint8_t* expected = buffer_data(&input_UTF8, expected_return);
 		ASSERT_NE(nullptr, expected) <<
